Read CSV data

+++ App/Server/Models/TimeSeries.cs
@@ -4,7 +4,7 @@ namespace AspCoreServer.Models
     public class TimeSeries
     {
         public string Name { set; get; }
-        public float[][] Data { set; get; }
+        public double[][] Data { set; get; }
 
         //Setting Default value
         public TimeSeries()
